backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
anthropic==0.39.0
networkx==3.2.1
pypdf==4.0.1
python-docx==1.1.0
supabase==2.3.4
httpx==0.25.2
pydantic
numpy
pandas
openpyxl
